requirements.txt
asgiref==3.2.2
asn1crypto==1.0.1
attrs==19.2.0
autobahn==19.10.1
Automat==0.7.0
certifi==2019.9.11
cffi==1.12.3
channels==2.3.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
Django==2.2.6
django-annoying==0.10.5
django-cors-headers==3.1.1
django-eventstream==2.6.0
django-filter==2.2.0
django-grip==2.0.1
django-regex-field==1.3.1
djangorestframework==3.10.3
gripcontrol==3.3.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
pubcontrol==2.4.2
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
urllib3==1.25.6
Werkzeug==0.16.0
zope.interface==4.6.0
